chore: remove potentially redundant groups
`mutually:`, `generated:`, `matching:`, and `instantiated:` have
been removed since they, at the time, seem to be redundant and
make things more complicated than necessary.
