+++ src/test/resources/goldens/chalktalk/sources/handles_sources_with_a_homepage/phase2-structure.txt
@@ -2,7 +2,7 @@ Document(
   groups = [
              ResourceGroup(
                id = "id"
-               sourceSection = ResourceSection(
+               resourceSection = ResourceSection(
                  items = [
                            StringSectionGroup(
                              section = StringSection(
