+++ src/test/resources/mathlingua.math
@@ -33,37 +33,6 @@ maps:
 written: "\textrm{some.function}(x?)"
 
 
-[\integers]
-Defines: Z := (X, +)
-instantiated:
-. 'X := \set'
-. 'x + y := x \integer.+/ y'
-written: "\mathbb{Z}"
-
-
-[\naturals]
-Defines: N
-means: 'N is \type'
-generated:
-. inductively:
-  from:
-  . constant: 0
-  . constructor: succ(x)
-    from: '\natural'
-written: "\mathbb{N}"
-
-
-[\naturals2]
-Defines: N
-generated:
-. inductively:
-  from:
-  . constant: 0
-  . constructor: succ(x)
-    from: '\natural'
-written: "\mathbb{N}"
-
-
 [\even.primes]
 Defines: X
 means: 'X is \set'
@@ -96,23 +65,6 @@ written: "f1(x?)"
 Defines: f(x)
 evaluated: 'f(x) := x + 1'
 written: "f1(x?)"
-
-
-[\f2(x)]
-Defines: f(x)
-means: 'f is \function'
-evaluated:
-. matching:
-  . 'f(x?) ::= x + 1'
-written: "f2(x?)"
-
-
-[\f22(x)]
-Defines: f(x)
-evaluated:
-. matching:
-  . 'f(x?) ::= x + 1'
-written: "f2(x?)"
 
 
 [\f3(x)]
@@ -181,19 +133,6 @@ Foundation:
     all: 'x'
     suchThat: 'x'
   written: "X"
-
-
-Mutually:
-. [\X]
-  Defines: X
-  means: 'X is \X'
-  satisfying: 'X is \Y'
-  written: "X?"
-. [\Y]
-  Defines: Y
-  means: 'Y is \Y'
-  satisfying: 'Y is \X'
-  written: "Y?"
 
 
 [x + y]
